Add SD and DAC functions, clean up
This changeset attempts to add better separation between the STM32 hardware setup and the "HAL" itself.

This includes moving things out of main.c, although it currently includes the loading and playing of a file from SD card until this is formalised into the audio API.

blit::debug has been defined as a simple method that prints a message to the top left corner- this should avoid it being a nullptr trap.

Updating the backlight has been added alongside the other LED updates. I'd like to bring the HAL API functions that main.c needs to call down to a bare minimum.

blit::joystick.y is now being inverted so that stick movement matches screen coordinates.

blit::battery has been converted to a float and will report the battery voltage in volts.

Various tweaks and tidyup have been applied to hardware-test

+++ 32blit/engine/input.cpp
@@ -10,7 +10,7 @@ namespace blit {
   vec3 tilt;
   float hack_left;
   float hack_right;
-  int battery;
+  float battery;
 
   /**
    * Return pressed state of a button or buttons.
